Portability and signal management improvements
Ensure most of the codebase is as POSIX as possible.
Some features such as the whole <sys/reboot.h> is of course not POSIX,
but some functionalities wouldn't be possible without them, so they're an
acceptable compromise on feature-set/compatibility balance.

Also reviewed the codebase to ensure all opened file descriptors
are set with the close-on-exec flag, so the daemon start routine
wouldn't rely on specific mechanisms to close opened file descriptors.

Added a fix in the daemon configuration parser,
to avoid interpreting badly named section.

Completely reworked signal handling and the main loop.
Signal handlers are now just notify being triggered through
global variables, and it is the main loop's responsibility
to execute the right task, being reboot, configuration reload
or child reaping. This allowed moving the child reaping logic
in the main.c and simplify both rc parsing and the teardown
sequence. The teardown sequence is now also based on SIGALRM
and doesn't use nanosleep anymore.

* .gitignore: Added artifacts.
* README.md: Modernize descriptions.
* bconf: Added new options, removed procfs options.
* bconf.mk: Updated for options and remove clover dependency.
* docs/cyberd.*.md: Corrected errors.
* docs/endpoints.md: Made description more formal.
* src/.gitignore: Remove 'cyberd' rule for source directory.
* src/cyberctl/main.c: Renamed, improved portability.
* src/cyberd/capabilities.h: Ensure more type safety.
* src/cyberd/configuration.c: Improved readability.
* src/cyberd/configuration.h: Renamed teardown in cleanup.
* src/cyberd/daemon.c: Improve portability by removing fd closing.
* src/cyberd/daemon.h: Removed running state in started state.
* src/cyberd/daemon_conf.c: Added optional NSIG, inclusive max uid/gid.
* src/cyberd/daemon_conf.h: Aligned bitfields types definitions.
* src/cyberd/main.c: Major rework, see above.
* src/cyberd/signals.[ch]: Simplification, see above.
* src/cyberd/socket_connection_node.c: Can parse up to NAME_MAX + 1.
* src/cyberd/socket_endpoint_node.c: Added close-on-exec on accepted sockets.
* src/cyberd/socket_switch.[ch]: Socket switch only modifies readfds now.
* src/cyberd/spawns.[ch]: Accomodate new teardown logic.
* src/cyberd/tree.[ch]: Coding-style + added assert to simplify static analysis.
* test/.gitignore: Ignore test binaries.
* test/cyberd-tree.c: Removed dependency on cover.